Boost: retire the rsm_jetpack_ui_modernization_boost gate + legacy SettingsPage
The wp-build chassis (Overview + Settings tabs) is now the only admin
experience for Boost — the modernization filter, the legacy
`render_settings` callback, and the bespoke `SettingsPage` shell with
its `Tips` + `Support` chrome are all retired.

PHP:

* Remove the `MODERNIZATION_FILTER` constant + `is_modernized()` helper
  on `Admin`. `handle_admin_menu` always wires the chassis render
  function; `maybe_load_wp_build` always loads `build/build.php`;
  `enqueue_scripts` always localizes the chassis prereqs handle.
* Drop `render_settings()` and the `<div id="jb-admin-settings">`
  fallback markup. The dual-bundle architecture (legacy webpack ↔
  chassis settings tab mount) is preserved.
* `wp-js-data-sync.php` always attaches the data-sync client to
  `jetpack-boost-dashboard-wp-admin-prerequisites`.

JS:

* `main.tsx` deleted; its route definitions and `Tracks` page-view
  wrapper merged into `index.tsx` (chassis-only entry).
* `index.tsx` drops the `#jb-admin-settings` fallback branch and the
  `LEGACY_ROOT_ID` constant; the bundle now only mounts inside the
  chassis's `#jb-settings-tab-mount`.
* Whole `layout/settings-page/` directory removed (SettingsPage + Tips
  + Support). The legacy `#jb-admin-settings` background rule in
  `wp-admin.scss` goes with it.

Dropping the legacy webpack pipeline + adding the `jetpack_boost_tab_view`
Tracks event are deferred follow-ups — the legacy bundle still owns
Settings tab content because the browser-side Critical CSS generator
depends on Node-only modules wp-build's esbuild can't polyfill.
