+++ projects/plugins/boost/app/assets/src/css/main/wp-admin.scss
@@ -87,10 +87,6 @@ p {
 	min-height: 100vh;
 }
 
-.jetpack_page_jetpack-boost #jb-admin-settings {
-	background: #fff;
-}
-
 .jetpack_page_jetpack-boost #dolly {
 	background: #fff;
 }
