+++ projects/plugins/boost/app/assets/src/js/index.tsx
@@ -1,55 +1,125 @@
 import { DataSyncProvider } from '@automattic/jetpack-react-data-sync-client';
 import * as WPElement from '@wordpress/element';
-import { StrictMode } from 'react';
-import { createHashRouter } from 'react-router';
+import { StrictMode, useEffect } from 'react';
+import { createHashRouter, redirect, useLocation } from 'react-router';
 import { RouterProvider } from 'react-router/dom';
+import type { JSX } from 'react';
 import CriticalCssProvider from '$features/critical-css/critical-css-context/critical-css-context-provider';
 import { NoticeProvider } from '$features/notice/context';
 import NoticeManager from '$features/notice/manager';
-import Main from './main';
+import { useGettingStarted } from '$lib/stores/getting-started';
+import { recordBoostEvent } from '$lib/utils/analytics';
 import Index from './pages/index';
 import AdvancedCriticalCss from './pages/critical-css-advanced/critical-css-advanced';
 import CacheDebugLog from './pages/cache-debug-log/cache-debug-log';
+import GettingStarted from './pages/getting-started/getting-started';
+import PurchaseSuccess from './pages/purchase-success/purchase-success';
 import '../css/admin-style.scss';
 
 const TAB_MOUNT_ID = 'jb-settings-tab-mount';
-const LEGACY_ROOT_ID = 'jb-admin-settings';
 const CHASSIS_APP_ID = 'jetpack-boost-dashboard-wp-admin-app';
 
-const tabMountRouter = createHashRouter( [
-	{
-		path: '*',
-		element: (
-			<NoticeProvider>
-				<CriticalCssProvider>
-					<Index />
-					<NoticeManager />
-				</CriticalCssProvider>
-			</NoticeProvider>
-		),
-	},
-	{
-		path: '/cache-debug-log',
-		element: <CacheDebugLog />,
-	},
-	{
-		path: '/critical-css-advanced',
-		element: (
-			<NoticeProvider>
-				<CriticalCssProvider>
-					<AdvancedCriticalCss />
-					<NoticeManager />
-				</CriticalCssProvider>
-			</NoticeProvider>
-		),
-	},
-] );
+/**
+ * Records `page_view*` Tracks events whenever the hash router moves
+ * between routes.
+ *
+ * @param props
+ * @param props.children - The actual page to render
+ */
+const Tracks = ( { children }: { children: JSX.Element } ) => {
+	const location = useLocation();
+
+	useEffect( () => {
+		let path = location.pathname.replace( /[-/]/g, '_' );
+		if ( path === '_' ) {
+			path = '_settings';
+		}
+
+		recordBoostEvent( `page_view${ path }`, {
+			path: location.pathname,
+		} );
+	}, [ location ] );
+
+	return children;
+};
+
+const useBoostRouter = () => {
+	const { shouldGetStarted } = useGettingStarted();
+
+	const checkForGettingStarted = () => {
+		if ( shouldGetStarted ) {
+			return redirect( '/getting-started' );
+		}
+		return null;
+	};
+
+	return createHashRouter( [
+		{
+			path: '*',
+			loader: checkForGettingStarted,
+			element: (
+				<NoticeProvider>
+					<CriticalCssProvider>
+						<Tracks>
+							<Index />
+						</Tracks>
+						<NoticeManager />
+					</CriticalCssProvider>
+				</NoticeProvider>
+			),
+		},
+		{
+			path: '/cache-debug-log',
+			loader: checkForGettingStarted,
+			element: (
+				<Tracks>
+					<CacheDebugLog />
+				</Tracks>
+			),
+		},
+		{
+			path: '/critical-css-advanced',
+			loader: checkForGettingStarted,
+			element: (
+				<NoticeProvider>
+					<CriticalCssProvider>
+						<Tracks>
+							<AdvancedCriticalCss />
+						</Tracks>
+						<NoticeManager />
+					</CriticalCssProvider>
+				</NoticeProvider>
+			),
+		},
+		{
+			path: '/getting-started',
+			element: (
+				<Tracks>
+					<GettingStarted />
+				</Tracks>
+			),
+		},
+		{
+			path: '/purchase-successful',
+			element: (
+				<Tracks>
+					<PurchaseSuccess />
+				</Tracks>
+			),
+		},
+	] );
+};
+
+function TabMountApp() {
+	const router = useBoostRouter();
+	return <RouterProvider router={ router } />;
+}
 
 const renderTabMount = ( target: HTMLElement ) => {
 	WPElement.createRoot( target ).render(
 		<StrictMode>
 			<DataSyncProvider>
-				<RouterProvider router={ tabMountRouter } />
+				<TabMountApp />
 			</DataSyncProvider>
 		</StrictMode>
 	);
@@ -60,6 +130,7 @@ const renderTabMount = ( target: HTMLElement ) => {
  * before mounting. The chassis renders the panel asynchronously after
  * its bundle hydrates, so the element doesn't exist when this script's
  * first pass runs at footer load.
+ *
  * @param done
  */
 const waitForTabMount = ( done: ( target: HTMLElement ) => void ) => {
@@ -81,33 +152,16 @@ const waitForTabMount = ( done: ( target: HTMLElement ) => void ) => {
 };
 
 /**
- * Initial render function.
- *
- * Three modes, picked at boot:
- *
- * 1. Modernized (chassis present, legacy root absent) — wait for the
- *    chassis's `#jb-settings-tab-mount` div to appear, then render a
- *    slim hash router (no chassis chrome) into it. Skips `SettingsPage`'s
- *    `BoostAdminPage`, `SpeedScore`, `Tips`, `Support` — those belong
- *    to the chassis. Hash routing coexists with the chassis's search-
- *    param routing (`?tab=...`); the two never collide.
- * 2. Legacy (`#jb-admin-settings` present) — falls through to the full
- *    `<Main />` (hash router + SettingsPage chrome) when the
- *    modernization filter is off.
- * 3. Neither root — no-op (page isn't ours).
+ * Boot the legacy webpack bundle. The wp-build chassis owns the page chrome
+ * + Overview tab; this entry waits for the chassis's
+ * `#jb-settings-tab-mount` div, then renders a slim hash router (no chassis
+ * chrome) into it. Hash routing coexists with the chassis's search-param
+ * routing (`?tab=...`); the two never collide.
  */
 function render() {
-	const legacyContainer = document.getElementById( LEGACY_ROOT_ID );
-	if ( legacyContainer !== null ) {
-		WPElement.createRoot( legacyContainer ).render( <Main /> );
+	if ( document.getElementById( CHASSIS_APP_ID ) === null ) {
 		return;
 	}
-
-	const isChassisPage = document.getElementById( CHASSIS_APP_ID );
-	if ( isChassisPage === null ) {
-		return;
-	}
-
 	waitForTabMount( renderTabMount );
 }
