--- projects/plugins/boost/app/assets/src/js/layout/settings-page/settings-page.module.scss
@@ -1,3 +1,0 @@
-:global(.jb-section).section {
-	z-index: 14;
-}
