--- projects/plugins/boost/app/assets/src/js/layout/settings-page/settings-page.tsx
@@ -1,48 +1,0 @@
-import SpeedScore from '$features/speed-score/speed-score';
-import BoostAdminPage from '$layout/boost-admin-page/boost-admin-page';
-import Support from './support/support';
-import Tips from './tips/tips';
-import clsx from 'clsx';
-import styles from './settings-page.module.scss';
-import { usePremiumFeatures } from '$lib/stores/premium-features';
-import CriticalCssProvider from '$features/critical-css/critical-css-context/critical-css-context-provider';
-import NoticeManager from '$features/notice/manager';
-import { NoticeProvider } from '$features/notice/context';
-import type { ReactNode } from 'react';
-
-type SettingsPageProps = {
-	children: ReactNode;
-};
-
-const SettingsPage = ( { children }: SettingsPageProps ) => {
-	const premiumFeatures = usePremiumFeatures();
-	const hasPrioritySupport = premiumFeatures && premiumFeatures.includes( 'support' );
-
-	return (
-		<NoticeProvider>
-			<CriticalCssProvider>
-				<BoostAdminPage>
-					<div id="jb-dashboard" className="jb-dashboard jb-dashboard--main">
-						<div className="jb-section jb-section--alt jb-section--scores">
-							<SpeedScore />
-						</div>
-
-						{ children && (
-							<div className={ clsx( 'jb-section jb-section--main', styles.section ) }>
-								{ children }
-							</div>
-						) }
-
-						<Tips />
-
-						{ hasPrioritySupport && <Support /> }
-
-						<NoticeManager />
-					</div>
-				</BoostAdminPage>
-			</CriticalCssProvider>
-		</NoticeProvider>
-	);
-};
-
-export default SettingsPage;
