--- projects/plugins/boost/app/assets/src/js/layout/settings-page/support/support.module.scss
@@ -1,13 +1,0 @@
-.support {
-	display: grid;
-	grid-template-columns: auto auto;
-	align-items: center;
-}
-
-.title {
-	margin-bottom: 24px !important; // @todo - use variable $halfling
-}
-
-.cta {
-	text-align: right;
-}
