--- projects/plugins/boost/app/assets/src/js/layout/settings-page/tips/tips.module.scss
@@ -1,57 +1,0 @@
-@use "$css/main/variables";
-@use "$css/main/mixins";
-
-.tips {
-	margin: 0 auto;
-
-	.tips-title {
-		margin-bottom: variables.$padding;
-	}
-
-	.tips-items {
-
-		@include mixins.breakpoint(md) {
-			display: grid;
-			grid-template-columns: 1fr 1fr;
-			gap: 2em;
-		}
-	}
-
-	.item {
-		margin-bottom: 2em;
-
-		@include mixins.breakpoint(sm) {
-			display: grid;
-			grid-template-columns: 80px 1fr;
-			gap: 1em;
-		}
-
-		.item-rate {
-			font-size: 24px;
-			font-weight: 400;
-			line-height: 30px;
-			color: variables.$primary-black;
-
-			@include mixins.breakpoint(md) {
-				flex-basis: 20%;
-				font-size: 36px;
-				line-height: 43px;
-			}
-		}
-
-		.item-description {
-			flex-basis: 70%;
-			font-weight: 400;
-			line-height: 24px;
-			color: variables.$primary-grey;
-
-			@include mixins.breakpoint(md) {
-				flex-basis: 80%;
-			}
-
-			a {
-				color: variables.$gray_90;
-			}
-		}
-	}
-}
